api/requirements.txt
django==3.0.1
djangorestframework
django-monthfield
django-polymorphic
django-rest-polymorphic
django-nested-admin